...
Lamonte Cranston once hired a new Chinese manservant.  While describing his
duties to the new man, Lamonte pointed to a bowl of candy on the coffee
table and warned him that he was not to take any.  Some days later, the new
manservant was cleaning up, with no one at home, and decided to sample some
of the candy.  Just than, Cranston walked in, spied the manservant at the
candy, and said:
	"Pardon me Choy, is that the Shadow's nugate you chew?"