...
Attorney General Edwin Meese III explained why the Supreme Court's Miranda
decision (holding that subjects have a right to remain silent and have a
lawyer present during questioning) is unnecessary: "You don't have many
suspects who are innocent of a crime.  That's contradictory.  If a person
is innocent of a crime, then he is not a suspect."
		-- U.S. News and World Report, 10/14/85